refactor(store home page): split server code from client code
