+++ frontend/templates/page/index.tsx
@@ -1,11 +1,6 @@
 import { Box } from '@chakra-ui/react';
-import { DEFAULT_STORE_CODE } from '@config/env';
-import { flags } from '@config/flags';
-import { ifixitOriginFromHost } from '@helpers/path-helpers';
 import { assertNever } from '@ifixit/helpers';
-import { DefaultLayout, getLayoutServerSideProps } from '@layouts/default';
-import { findPage } from '@models/page';
-import { GetServerSideProps } from 'next';
+import { DefaultLayout } from '@layouts/default';
 import {
    PageTemplateProps,
    usePageTemplateProps,
@@ -14,7 +9,7 @@ import { BrowseSection } from './sections/BrowseSection';
 import { HeroSection } from './sections/HeroSection';
 import { StatsSection } from './sections/StatsSection';
 
-export const PageTemplate: NextPageWithLayout<PageTemplateProps> = () => {
+const PageTemplate: NextPageWithLayout<PageTemplateProps> = () => {
    const { page } = usePageTemplateProps();
    return (
       <Box>
@@ -41,36 +36,4 @@ PageTemplate.getLayout = function getLayout(page, pageProps) {
    return <DefaultLayout {...pageProps.layoutProps}>{page}</DefaultLayout>;
 };
 
-export const getServerSideProps: GetServerSideProps<PageTemplateProps> = async (
-   context
-) => {
-   if (!flags.STORE_HOME_PAGE_ENABLED) {
-      return {
-         notFound: true,
-      };
-   }
-
-   const layoutProps = await getLayoutServerSideProps({
-      storeCode: DEFAULT_STORE_CODE,
-   });
-   const page = await findPage({
-      path: '/Store',
-   });
-
-   if (page == null) {
-      return {
-         notFound: true,
-      };
-   }
-
-   const pageProps: PageTemplateProps = {
-      layoutProps,
-      appProps: {
-         ifixitOrigin: ifixitOriginFromHost(context),
-      },
-      page,
-   };
-   return {
-      props: pageProps,
-   };
-};
+export default PageTemplate;
